% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginHost.R
\name{hostFunction}
\alias{hostFunction}
\title{Project workspace initializer}
\usage{
hostFunction(projectName)
}
\arguments{
\item{projectName}{A string}
}
\description{
\code{projectInit} sources the \code{00-init.R} or \code{projectInit.R} 
script for the project. You pass a complete folder or a relative path. This 
function is intended to be run just once per session.
}
